Move common NeXus stuff to ess.powder.nexus
